refactor: update transaction summary fields from Long to Double
This commit updates the data type of the fields in the `TransactionSummary` data class from `Long` to `Double`.

This change ensures floating-point precision for financial calculations and addresses potential type mismatches. The UI components in `ProfilerScreen`, `TransactionsScreen`, and the mock data in `ProfileDetailScreen` have been updated accordingly to handle `Double` values.

+++ android-compose/FinOps/app/src/main/java/com/example/fin_ops/data/remote/dto/ProfilerTransactionDto.kt
@@ -51,17 +51,17 @@ data class ProfilerTransactionDto(
 // Represents the summary block in the response
 data class TransactionSummary(
     @SerializedName("total_deposits")
-    val totalDeposits: Long,
+    val totalDeposits: Double,
     @SerializedName("total_withdrawals")
-    val totalWithdrawals: Long,
+    val totalWithdrawals: Double,
     @SerializedName("total_charges")
-    val totalCharges: Long,
+    val totalCharges: Double,
     @SerializedName("net_amount")
-    val netAmount: Long,
+    val netAmount: Double,
     @SerializedName("transaction_difference")
-    val transactionDifference: Long,
+    val transactionDifference: Double,
     @SerializedName("credit_uncountable")
-    val creditUncountable: Long
+    val creditUncountable: Double
 )
